The new chat library.
This is the new chat library.  Instead of using the express framework, I'm
now going to use the built in http/https server of node, to serve up
the raw javascript.  Along with the filesystem or fs node module, I'm
able to stich the socket.io-client with the chat libray before I serve
up the code to the client.  This proved to be a very simple solution to
meet our goal.